requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
botocore==1.15.32
celery==4.4.2
Django==3.0.4
django-extensions==2.2.9
django-mysql==3.3.0
djangorestframework==3.11.0
dnspython==1.16.0
docutils==0.15.2
flower==0.9.4
gunicorn==20.0.4
humanize==0.5.1
importlib-metadata==1.6.0
jmespath==0.9.5
kombu==4.6.8
mysqlclient==1.4.6
numpy==1.18.2
pandas==1.0.3
Pillow==7.0.0
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
